web-animations/keyframe-effect/keyframe-effect-normalizing-keyframes.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Test normalizing a sequence of keyframes</title>
<meta name="assert" content="For each call to setFrames or the KeyframeEffect constructor, normalization is performed on the passed in frames parameter before storing its value">
<link rel="help" href="http://w3c.github.io/web-animations/#normalizing-a-sequence-of-keyframes">
<link rel="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<link rel="author" title="Aleksei Yu. Semenov" href="mailto:a.semenov@unipro.ru">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
<body>
<div id="log"></div>
<script>
// Creates two effects that tested in the tests below. One with frames,
// defined by the constructor. And the second with keyframes defined by setFrames()
function createEffectsToTest(keyframes) {
    var effect1 = new KeyframeEffect(keyframes);
    var effect2 = new KeyframeEffect([]);
    effect2.setFrames(keyframes);
    return [effect1, effect2];
}

var effectsDescription = ['KeyframeEffect constructor', 'KeyframeEffect.setFrames()'];

// produces expected keyframe sequence from given single input keyframe
// or input sequence of keyframes
function expected(){
    var result = [];
    for (var i=0; i<arguments.length; i++){
        var a = arguments[i];
        if (!Array.isArray(a)){
            a = [a];
        }
        for (var k=0; k<a.length; k++) {
            var keyframe = {offset: null, easing: 'linear', composite: null};
            for (var propertyName in a[k]){
                keyframe[propertyName] = a[k][propertyName];
            }
            if (keyframe.offset !== null){
                keyframe.computedOffset = keyframe.offset;
            }
            result.push(keyframe);
        }
    }
    return result;
}


test(function() {
    var effects = createEffectsToTest([]);

    effects.forEach(function(effect, index) {
        assert_array_equals(effect.getFrames(), [], effectsDescription[index] +
            ' should initialize frames sequence with empty array');
    });
}, 'Test keyframes normalization of empty sequence of frames does nothing');

// 1. If frames is a single Keyframe object, replaces frames with a new sequence with a single item that is the previous value of frames.
test(function() {
    var keyframe = {
        top: '10px',
        offset: 0.5,
        easing: 'ease',
        composite: 'add'
    };
    var expectedFrames = expected(keyframe);
    var effects = createEffectsToTest(keyframe);
    effects.forEach(function(effect, index) {
        assert_array_equals(effect.getFrames(), expectedFrames,
            effectsDescription[index] + ' should replace input keyframe with the sequence');
    });
}, 'Test keyframes normalization replaces single input Keyframe with ' +
    'sequence of that single keyframe instance');

// 2. If frames is not loosely sorted by offset, then,
test(function() {
    var keyframes = [{top: '10px', offset: 0.1}, {top: '20px', offset: 0.2}];
    var effects = createEffectsToTest(keyframes);
    var expectedFrames = expected(keyframes);

    effects.forEach(function(effect, index) {
        assert_array_equals(effect.getFrames(), expectedFrames,
            effectsDescription[index] + ' should keep the order of the frames');
    });
}, 'Test keyframes normalization keeps the order of the frames, ' +
    'if input keyframe sequense is already loosely sorted by offset');

test(function() {
    var keyframes = [
        {top: '0px'},
        {top: '10px', offset: 0.1},
        {top: '20px'},
        {top: '30px', offset: 0.3},
        {top: '100px'}
    ];
    var effects = createEffectsToTest(keyframes);
    var expectedFrames = computeOffset(expected(keyframes), 'distribute');

    effects.forEach(function(effect, index) {
        assert_array_equals(effect.getFrames(), expectedFrames,
            effectsDescription[index] + ' should keep the order of the frames');
    });
}, 'Test keyframes normalization keeps the order of the frames, ' +
    'if input keyframe sequense is already loosely sorted by offset. ' +
    'Some keyframes have no offset defined');

// If each Keyframe has a non-null offset,
//    Sort frames in ascending order by offset.
test(function() {
    var keyframes = [
        {top: '20px', offset: 0.2},
        {top: '0px', offset: 0},
        {top: '10px', offset: 0.1},
        {top: '100px', offset: 1},
        {top: '80px', offset: 0.8}
    ];
    var effects = createEffectsToTest(keyframes);
    var expectedFrames = expected(
        keyframes[1],
        keyframes[2],
        keyframes[0],
        keyframes[4],
        keyframes[3]
    );
    effects.forEach(function(effect, index) {
        assert_array_equals(effect.getFrames(), expectedFrames,
            effectsDescription[index] + ' should loosely sort ' +
            'input sequence of keyframes by offset');
    });
}, 'Test keyframes normalization loosely sorts sequence of input keyframes by offset');

// Otherwise,
//    Throw a DOMException of type InvalidModificationError.
test(function() {
    var keyframes = [
        {top: '0px'},
        {top: '10px', offset: 0.3},
        {top: '20px'},
        {top: '30px', offset: 0.1},
        {top: '100px'}
    ];

    assert_throws('InvalidModificationError', function() {
        new KeyframeEffect(keyframes);
    }, 'KeyframeEffect constructor should throw InvalidModificationError, ' +
        'if input sequence is not sorted and some keyframes have no offset defined');
}, 'Test KeyframeEffect constructor throws InvalidModificationError, ' +
    'if input sequence is not sorted and some keyframes have no offset defined');

test(function() {
    var keyframes = [
        {top: '0px'},
        {top: '10px', offset: 0.3},
        {top: '20px'},
        {top: '30px', offset: 0.1},
        {top: '100px'}
    ];
    var effect = new KeyframeEffect([]);

    assert_throws('InvalidModificationError', function() {
        effect.setFrames(keyframes);
    }, 'KeyframeEffect.setFrames() should throw InvalidModificationError, ' +
        'if input sequence is not sorted and some keyframes have no offset defined');
}, 'Test KeyframeEffect.setFrames() throws InvalidModificationError, ' +
    'if input sequence is not sorted and some keyframes have no offset defined');

// 3. If there exist any Keyframe objects in frames whose offset member is non-null
// and less than zero, remove all keyframes objects from the start of frames up to
// and including the keyframe with the largest non-null offset that is still less than zero.
// 4. Likewise, if there exist any Keyframe objects in frames whose offset member
// is non-null and greater than one, remove all keyframes from the keyframe
// with the smallest non-null offset that is still greater than one until the end of frames.
test(function() {
    var keyframes = [
        {top: '20px', offset: 0.2},
        {top: '-10000px', offset: Number.MIN_VALUE},
        {top: '0px', offset: 0},
        {top: '10px', offset: 0.1},
        {top: '-10px', offset: -0.1},
        {top: '2340px', offset: 234},
        {top: '100px', offset: 1},
        {top: '120px', offset: 1.2},
        {top: '80px', offset: 0.8},
        {top: '-1000px', offset: -100},
        {top: '10000px', offset: Number.MAX_VALUE}
    ];
    var effects = createEffectsToTest(keyframes);
    var expectedFrames = expected(
        keyframes[2],
        keyframes[3],
        keyframes[0],
        keyframes[8],
        keyframes[6]
    );

    effects.forEach(function(effect, index) {
        assert_array_equals(effect.getFrames(), expectedFrames,
            effectsDescription[index] + ' should remove frames with offset ' +
                'value outside [0..1]');
    });
}, 'Test keyframes normalization removes keyframes with offsets greater than 1 or less than 0');

test(function() {
    var keyframes = [
        {top: '-10000px', offset: Number.MIN_VALUE},
        {top: '-1000px', offset: -100},
        {top: '-50px'},
        {top: '-10px', offset: -0.1},
        {top: '-5px'},
        {top: '10px', offset: 0.1},
        {top: '20px', offset: 0.2},
        {top: '50px'},
        {top: '80px', offset: 0.8},
        {top: '90px', offset: 0.9},
        {top: '95px'},
        {top: '100px'},
        {top: '120px', offset: 1.2},
        {top: '150px'},
        {top: '2340px', offset: 234},
        {top: '10000px', offset: Number.MAX_VALUE}
    ];
    var effects = createEffectsToTest(keyframes);
    var expectedFrames = computeOffset(expected(keyframes.slice(4, 12)), 'distribute');

    effects.forEach(function(effect, index) {
        assert_array_equals(effect.getFrames(), expectedFrames,
            effectsDescription[index] + ' should remove frames with offset ' +
                'value outside [0..1]');
    });
}, 'Test keyframes normalization removes keyframes with offsets greater than 1 or less than 0. ' +
    'Keyframes have null and non-null offsets and are loosely sorted by offset');

test(function() {
    var keyframe = {top: '10px', someUnsupportedProperty: 'someValue'};
    var effects = createEffectsToTest(keyframe);
    var expectedFrames = expected({top: '10px', computedOffset: 0});

    effects.forEach(function(effect, index) {
        assert_array_equals(effect.getFrames(), expectedFrames,
            effectsDescription[index] + ' should remove not supported properties');
    });
}, 'Test keyframes normalization of single Keyframe removes properties not supported ' +
    'by implementation');

test(function() {
    var keyframes = [
        {top: '10px', someUnsupportedProperty: 'someValue', offset: 0.1},
        {top: '20px', someUnsupportedProperty: 'someValue', offset: 0.2}
    ];
    var effects = createEffectsToTest(keyframes);
    var expectedFrames = expected(keyframes);

    effects.forEach(function(effect, index) {
        assert_array_equals(effect.getFrames(), expectedFrames,
            effectsDescription[index] + ' should remove not supported properties');
    });
}, 'Test keyframes normalization of the keyframe sequence removes properties ' +
    'not supported by implementation');

// TODO Add test of keyframes with the same offset
</script>
</body>
